Export IntegersCanFit so it can be used from R or other bindings
